moving functionality to PulsarHelper
